audit-profile-row-reorder: move reorderRows off the public barrel
